requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.0.4
apache-airflow-providers-snowflake==5.4.0
apache-airflow-providers-slack==8.6.2
apache-airflow-providers-common-sql==1.12.0